requirements.txt
django
django-extensions
django-sesame
djangorestframework
numpy
matplotlib
selenium

